move job execution to dedicated thread
