+++ canisters/chain_fusion/src/logs.rs
@@ -1,5 +1,3 @@
-use ic_cdk::api;
-
 use std::{
     cmp::{min, Ordering},
     ops::{Add, Div, Sub},
@@ -15,7 +13,6 @@ use ic_cdk::println;
 
 use crate::{
     guard::TimerGuard,
-    job::execute_job,
     job::job,
     state::{mutate_state, read_state, State, TaskType},
 };
@@ -156,21 +153,6 @@ pub async fn scrape_eth_logs() {
                     return;
                 }
             };
-    }
-
-    // if we reach this point, we have successfully scraped all the logs up to the last observed block
-    // we can now check if there are any jobs to execute
-    let current_timestamp = api::time() / 1_000_000_000; // converted to seconds
-    let earliest_job = read_state(|s| s.get_earliest_job());
-    if let Some((job_id, job_execution_time)) = earliest_job {
-        if job_execution_time <= current_timestamp {
-            ic_cdk::println!("Executing job with ID: {:?}", job_id);
-            execute_job(job_id).await;
-            mutate_state(|s| s.remove_job(&job_id)); // remove the job from the queue
-            ic_cdk::println!("Executed job with ID: {:?}", job_id);
-        }
-    } else {
-        ic_cdk::println!("No jobs to execute");
     }
 }
